requirements.txt
openai>=1.0.0
pillow>=10.0.0
transformers>=4.30.0
tiktoken>=0.5.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
matplotlib>=3.7.0
pytest>=7.0.0
